Move lightTypes lookup table from high level component layer to mid level scene / resource layer.
